+++ src/FreecraftCore.Serializer.Stream.Default/Reader/DefaultStreamReaderStrategyAsync.cs
@@ -1,0 +1,108 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Text;
+using JetBrains.Annotations;
+
+#if !NET35
+using System.Threading.Tasks;
+#endif
+
+
+namespace FreecraftCore.Serializer
+{
+	/// <summary>
+	/// Default async implementation of the <see cref="IWireStreamReaderStrategy"/> that reads bytes from
+	/// an internally managed stream.
+	/// </summary>
+	public class DefaultStreamReaderStrategyAsync : DefaultStreamReaderStrategy, IWireStreamReaderStrategyAsync
+	{
+		/// <inheritdoc />
+		public DefaultStreamReaderStrategyAsync([NotNull] byte[] bytes) 
+			: base(bytes)
+		{
+
+		}
+
+		/// <inheritdoc />
+		public DefaultStreamReaderStrategyAsync([NotNull] Stream stream) 
+			: base(stream)
+		{
+
+		}
+
+		/// <inheritdoc />
+		public Task<byte> ReadByteAsync()
+		{
+			byte[] singleByteBuffer = new byte[1];
+
+			//would be -1 if it's invalid
+			Task<int> readResult = ManagedStream.ReadAsync(singleByteBuffer, 0, 1);
+
+			return new Task<byte>(() =>
+			{
+				int resultCount = readResult.Result;
+
+				if(resultCount == 0)
+					throw new InvalidOperationException($"Failed to read byte in {nameof(ReadByteAsync)}.");
+
+				return singleByteBuffer[0];
+			});
+		}
+
+		/// <inheritdoc />
+		public Task<byte> PeekByteAsync()
+		{
+			Task<byte> b = ReadByteAsync();
+
+			return new Task<byte>(() =>
+			{
+				byte resultByte = b.Result;
+
+				//Move it back one
+				ManagedStream.Position = ManagedStream.Position - 1;
+				return resultByte;
+			});
+		}
+
+		/// <inheritdoc />
+		public Task<byte[]> ReadAllBytesAsync()
+		{
+			return ReadBytesAsync((int)(ManagedStream.Length - ManagedStream.Position));
+		}
+
+		/// <inheritdoc />
+		public Task<byte[]> ReadBytesAsync(int count)
+		{
+			
+			byte[] bytes = new byte[count];
+
+			Task<int> readTask = ManagedStream.ReadAsync(bytes, 0, count);
+
+			return new Task<byte[]>(() =>
+			{
+				int resultCount = readTask.Result;
+
+				if (resultCount != count)
+					throw new InvalidOperationException($"Failed to read {count} bytes from the stream.");
+
+				return bytes;
+			});
+		}
+
+		/// <inheritdoc />
+		public Task<byte[]> PeakBytesAsync(int count)
+		{
+			Task<byte[]> bytes = ReadBytesAsync(count);
+
+			return new Task<byte[]>(() =>
+			{
+				//Now move the stream back
+				ManagedStream.Position = ManagedStream.Position - count;
+
+				return bytes.Result;
+			});
+		}
+	}
+}
